requirements.txt
behave~=1.2.6
nose>=1.3.7
requests>=2.10.0
trafaret>=0.7.1
jpath>=1.5
PyYAML~=6.0
setuptools~=60.2.0
behave-html-formatter~=0.9.10
